# doc/openapi/openapi-iam-v1.yaml
openapi: 3.0.3
info:
  title: IAM Service API
  description: |
    Multi-Tenant Identity and Access Management (IAM) Service API.

    ## Overview
    This API provides authentication, authorization, user management, participant
    management, member management, and masterdata capabilities for multi-tenant applications.

    ## Authentication
    Most endpoints require a JWT Bearer token in the `Authorization: Bearer <token>` header.
    Access tokens are obtained via the OTP-based login flow:
    1. `POST /api/v1/iam/login` — Initiate login, receive session ID
    2. `POST /api/v1/iam/login/{id}/verify-otp` — Verify OTP, receive access + refresh tokens

    Registration tokens (short-lived, HS256) are returned after OTP verification during
    the registration flow and must be used in subsequent registration steps.

    ## Multi-Tenancy
    Product-scoped endpoints (Participants, Members) require:
    - `Authorization: Bearer <access_token>` — JWT with tenant/product membership
    - `X-Tenant-ID: <uuid>` — Identifies the tenant context

    ## Rate Limiting
    Registration and login initiation endpoints are rate-limited to 10 req/min.
    Refresh token endpoint is rate-limited to 20 req/min.

    ## Error Codes
    | Code | HTTP | Meaning |
    |------|------|---------|
    | ERR_INTERNAL | 500 | Unexpected server error |
    | ERR_VALIDATION | 422 | Request validation failed |
    | ERR_NOT_FOUND | 404 | Resource not found |
...
    | ERR_FORBIDDEN | 403 | Access denied |
...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/iam/registrations:
    post:
      tags: [Registration]
      summary: Initiate registration
      description: |
        Starts the registration flow. Sends an OTP to the provided email address.
        Rate limited to 10 req/min.
      operationId: initiateRegistration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiateRegistrationRequest'
            example:
              email: user@example.com
      responses:
        '201':
          description: Registration initiated — OTP sent to email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiateRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/iam/registrations/{id}/verify-otp:
    post:
      tags: [Registration]
      summary: Verify registration OTP
      description: |
        Verifies the OTP sent during registration initiation.
        Returns a short-lived registration token for subsequent steps.
      operationId: verifyRegistrationOTP
      security: []
      parameters:
        - $ref: '#/components/parameters/RegistrationSessionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRegistrationOTPRequest'
            example:
              email: user@example.com
              otp_code: "123456"
      responses:
        '200':
          description: OTP verified — registration token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyRegistrationOTPResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/iam/registrations/{id}/resend-otp:
    post:
      tags: [Registration]
      summary: Resend registration OTP
      description: Resends the OTP for an existing registration session.
      operationId: resendRegistrationOTP
      security: []
      parameters:
        - $ref: '#/components/parameters/RegistrationSessionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendRegistrationOTPRequest'
            example:
              email: user@example.com
      responses:
        '200':
          description: OTP resent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/iam/registrations/{id}/status:
    get:
      tags: [Registration]
      summary: Get registration status
      description: Returns the current state of a registration session.
      operationId: getRegistrationStatus
      security: []
      parameters:
        - $ref: '#/components/parameters/RegistrationSessionID'
      responses:
        '200':
          description: Registration status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/iam/registrations/{id}/complete:
    post:
      tags: [Registration]
      summary: Complete registration (alternate flow)
      description: |
        Alternative completion endpoint that accepts password + basic profile in one step.
        Requires the registration token in the Authorization header.
      operationId: completeRegistration
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RegistrationSessionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteRegistrationRequest'
      responses:
        '201':
          description: Registration completed — tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteProfileRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # LOGIN
  # ==========================================
...
  /api/v1/iam/auth/logout-all:
    post:
      tags: [Auth]
      summary: Logout all sessions
      description: |
        Invalidates all tokens issued before this timestamp for the current user.
        Uses a user-level blacklist timestamp. All existing access and refresh tokens
        become invalid immediately.
      operationId: logoutAll
      responses:
        '200':
          description: All sessions terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # USERS
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token. Obtain via:
        - Login flow: POST /api/v1/iam/login → POST /api/v1/iam/login/{id}/verify-otp
        - Registration flow: POST /api/v1/iam/registrations → verify-otp → complete-profile
  parameters:
    RegistrationSessionID:
      name: id
      in: path
      required: true
      description: Registration session UUID
      schema:
        type: string
        format: uuid
      example: 018f3a2b-1c2d-7e4f-8a5b-6c7d8e9f0a1b
...
  responses:
    BadRequest:
      description: Bad request — malformed or missing parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "invalid request body"
            code: ERR_BAD_REQUEST
    Unauthorized:
      description: Unauthorized — missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "authentication required"
            code: ERR_UNAUTHORIZED
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "resource not found"
            code: ERR_NOT_FOUND
    Conflict:
      description: Conflict — resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "resource already exists"
            code: ERR_CONFLICT
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            error: "validation failed"
            code: ERR_VALIDATION
            errors:
              - field: email
                message: "must be a valid email address"
              - field: password
                message: "must be at least 8 characters"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "rate limit exceeded"
            code: ERR_TOO_MANY_REQUESTS
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "internal server error"
            code: ERR_INTERNAL
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
          example: "resource not found"
        code:
          type: string
          description: Machine-readable error code
          example: ERR_NOT_FOUND
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "validation failed"
        code:
          type: string
          example: ERR_VALIDATION
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: "must be a valid email address"
...
    SuccessMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "operation completed successfully"

    # ---- Health ----
...
    InitiateRegistrationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: user@example.com
    InitiateRegistrationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            registration_id:
              type: string
              format: uuid
              description: Session ID to use in subsequent registration steps
              example: 018f3a2b-1c2d-7e4f-8a5b-6c7d8e9f0a1b
            email:
              type: string
              format: email
              example: user@example.com
            expires_at:
              type: string
              format: date-time
            message:
              type: string
              example: "OTP sent to your email"
    VerifyRegistrationOTPRequest:
      type: object
      required: [email, otp_code]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        otp_code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: '^\d{6}$'
          example: "123456"
    VerifyRegistrationOTPResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            registration_token:
              type: string
              description: Short-lived token (HS256) for subsequent registration steps
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            token_expires_at:
              type: string
              format: date-time
            next_step:
              type: string
              enum: [SET_PASSWORD, COMPLETE_PROFILE]
              example: SET_PASSWORD
    ResendRegistrationOTPRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    RegistrationStatusResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            status:
              type: string
              enum: [PENDING_VERIFICATION, VERIFIED, PASSWORD_SET, COMPLETED]
              example: VERIFIED
            next_step:
              type: string
              example: SET_PASSWORD
...
    CompleteProfileRegistrationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            access_token:
              type: string
              example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
            refresh_token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              description: Access token lifetime in seconds
              example: 3600
    CompleteRegistrationRequest:
      type: object
      properties:
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        password_confirmation:
          type: string
          format: password
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        phone_number:
          type: string
          example: "+6281234567890"

    # ---- Login ----
...